+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 ### New features
 
 - String literals that have been stored with a locale specified (e.g. `nl-nl` for strings in Dutch) can now be accessed through `TripleSubject.getLocaleString` and `TripleSubject.getAllLocaleStrings`. Note that this will not return strings without a locale specified, nor will it return strings whose locale partially matches (e.g. strings with the locale `nl` will not be returned when `nl-nl` is requested, nor vice versa). As was already the case, `getString()` and `getAllStrings()` will not return strings with a locale specified.
+- To manipulate locale-specific strings, the `addLocaleString`, `setLocaleString` and `removeLocaleString` methods have been added to `TripleSubject`s.
 - Data type-specific methods have been added to `TripleSubject`s for manipulating Literals, to replace `addLiteral`, `setLiteral` and `removeLiteral`.
 
 ### Bugs fixed